pyproject.toml
[tool.poetry]
name = "model_deployment_operator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
kopf = { version="^1.37.2", extras=["uvloop"]}
kubernetes = "^29.0.0"
fire = "^0.6.0"
tritonclient = {extras = ["http"], version = "^2.45.0"}
boto3 = "^1.34.98"
boto3-stubs = {extras = ["s3"], version = "^1.34.98"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
testcontainers = "^4.3.0"
pytest-datadir = "^1.5.0"
triton-testcontainer = {git = "https://github.com/ogvalt/triton-testcontainer.git", rev = "v0.5.0"}
kubecrd = {git = "https://github.com/ogvalt/kubecrd.git", rev = "v0.5.1"}
pydantic = "^2.7.1"
jsonref = "^1.1.0"
pyyaml = "^6.0.1"
kubernetes-validate = "^1.29.1"
datamodel-code-generator = "^0.25.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = false
log_cli_level = "info"
pythonpath = [
  "."
]
